Move beers and breweries + unvalidated checkins to separate modules
