+++ beers/models/__init__.py
@@ -1,9 +1,11 @@
 from beers.models.models import (
-    Beer, Brewery, Contest, Contest_Beer, Contest_Bonus, Contest_Brewery,
-    Contest_Checkin, Contest_Player, Unvalidated_Checkin
+    Contest, Contest_Beer, Contest_Bonus, Contest_Brewery,
+    Contest_Checkin, Contest_Player
 )
 
 from beers.models.player import Player
+from beers.models.drinks import Beer, Brewery
+from beers.models.checkin import Unvalidated_Checkin
 
 __all__ = [
     'Beer', 'Brewery', 'Contest', 'Contest_Beer', 'Contest_Bonus', 'Contest_Brewery',
